LPS-108722 extract DepotItemSelectorViewRenderer to a class
